Start job rework to allow testing of the job itself
A best practice of creating CI/CD jobs is to have a test for the job
itself. That normally includes job execution with some stubs and fake
inventory. Luckily in Zuul it is possible to execute job again the
runtime job inventory allowing us to pre-test how the playbooks work.
We are not going to be able to send reports to DoJo, but testing all the
other parts is going to be very helpful.

Signed-off-by: Artem Goncharov <artem.goncharov@gmail.com>